requirements.txt
# =============================================================================
# SIEM ETL Pipeline — Python Dependencies
# =============================================================================

# --- Kafka ---
kafka-python==2.0.2

# --- Apache Spark (PySpark) ---
pyspark==3.5.0

# --- Apache Airflow ---
apache-airflow==2.7.3

# --- Elasticsearch ---
elasticsearch==8.11.0

# --- GeoIP enrichment ---
geoip2==4.7.0
maxminddb==2.5.1

# --- Structured logging ---
structlog==23.2.0

# --- Data quality ---
great-expectations==0.18.8

# --- HTTP / utilities ---
requests==2.31.0
python-dotenv==1.0.0
tenacity==8.2.3
faker==20.1.0

# --- Testing ---
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0

# --- Data processing ---
pandas==2.1.4
pyarrow==14.0.1

# --- MinIO / S3 ---
boto3==1.34.0
minio==7.2.0
